Added speaking capability & added Animal() function
I've been nagged to get the stories the story
generator made to be read out loud. I did and
it worked!

Also, there was this one part in
location_specific() that had been giving me a
headache, still giving me key errors for who
knows why. I took that part and put it in
its own function: Animal(). That's the last
of that problem.