Fixed menu bug in CaDPV2
There was a TypeError bug in the end-of-round menu in Council at
Daybreak proclaimer version 2.

I didn't test if it works, but I feel in my gut that it works!

For some reason, I get the feeling that the cards aren't being
shuffled. I'll go find the root cause of the problem if there is a
problem with the shuffling.