Removed debug lines
When I tested the shuffling code, I was unlucky. The output of the
random.shuffle() method returned the exact same list as the
unshuffled list. So I put a print(cards) command before and after
the random.shuffle(cards) line. I tested the new feature, and it's
good to go!